Changes in JRuby 9k when handling Time
Becore JRuby 9K, the time instance was using Joda time, this was forcing
the time to only show milliseconds precision, with the change to 9K they
switch their implementation to java.time which support nanoseconds
precision. But if you are using DateTime, it will still use joda time
and be limited to milliseconds precision.

Since internally we also use joda time we are also limited with
milliseconds precision, to support nanoseconds we will need to refactor
our timestamp class.

This commits change the expectation to only expect milliseconds
precision.